refactor: Separate tests for TreeStructs into different files.
